feat: 设备数据管理模块完成

+++ src/main/java/cn/edu/sjtu/ist/ecssbackendedge/entity/dto/Response.java
@@ -2,13 +2,13 @@ package cn.edu.sjtu.ist.ecssbackendedge.entity.dto;
 
 public class Response {
 
-    private Long code;
+    private Integer code;
 
     private String message;
 
     private Object object;
 
-    public Response(Long code, String message, Object object) {
+    public Response(Integer code, String message, Object object) {
         this.code = code;
         this.message = message;
         this.object = object;
